Delivery_Prediction/requirements.txt
snowflake-connector-python
pandas
numpy
scikit-learn
xgboost
matplotlib
seaborn
flask
pytest
mlflow